consumer/requirements.txt
psycopg2-binary==2.9.10
kafka-python==2.2.15
